Projeto_Final/requirements.txt
fastapi==0.111.0
psycopg2-binary==2.9.10
uvicorn==0.30.0
